requirements.txt
attrs==26.1.0
certifi==2026.4.22
charset-normalizer==3.4.7
idna==3.13
isodate==0.7.2
lxml==6.1.0
onvif_zeep==0.2.12
platformdirs==4.9.6
pytz==2026.2
requests==2.33.1
requests-file==3.0.1
requests-toolbelt==1.0.0
urllib3==2.7.0
zeep==4.3.2
